refactor the crawlers into one
This does remove the ability to "guess" mangareader images but I was not
using that anyways.  I'll find some way to add that back in the future.

The new crawler basically only has one entry point, handleManga.  If
it's a chapter that you're downloading a rule is added that only allows
tat chapter to go through, blocking everything else.  This simplifies
handleChapter.