add support for metadata formats
Right now we generate the ComicRack format and the CoMet format.  The
first seems widely supported by almost every app and the second just had
a proper definition and spec, unlike the other one.

Some fields are still not really supported (like the language) but I
think the important parts are there.